Ported the xfabric price code. Lot's of unnecessary debus still remain. Need to remove later
Former-commit-id: 79d7a840cbc962818327c7010692be1c3354fca2
Former-commit-id: 67e3b4a17ecae005df16cdef41626ab93f50ca4a [formerly 62899f8457e50b2641c8e9c06c7fd82c9f61757a] [formerly d19867a9136ddbdbf00af9b660b6c1490b644f7c [formerly 8340e8a846991a0cb7c49290c7b32d86d8637ec7]]
Former-commit-id: 48260d1c6d4f23d4730eaf4d8ba1ada18bbce597 [formerly 08438291fc2ce24f732d09a52e567d43bc24286c]
Former-commit-id: 84ac509e9698aa78c1fb9d0d9719c6314b5c14fa

+++ src/internet/model/tcp-newreno.h
@@ -75,7 +75,7 @@ protected:
   double getFlowIdealRate(std::string fkey);
   SequenceNumber32 ecn_highest;
   SequenceNumber32 last_outstanding_num;
-  uint32_t d0, dt;
+  double d0, dt;
 
   // Implementing ns3::TcpSocket -- Attribute get/set
   virtual void     SetSegSize (uint32_t size);
